Refactored AJAX and non-AJAX rendering of notices
Closes #589

Breaking Change: The div ID for warning notices
changes from lift__noticesContainer___warn to
lift__noticesContainer___warning to make it consistent
with the error and notice div IDs.

Inconsistent (and duplicated) code paths for AJAX
and non-AJAX rendering of notices resulted in problems
for end users. Refactored to remove a lot of duplicate
code, reorganized and documented a lot of helpers, and
changed NoticeType to an object instead of an enum
to allow for a richer functionality.