Clean up and refactor MapperSpec to run under SBT
Closes #927

Had to refactor quite a bit of the setup code to ensure that
it was being run. The previous incarnation relied on JUnit's
ordering guarantees, which didn't appear to hold under SBT.
Among other things, the MapperRules object appeared to actually
have multiple instances, as some runs failed because it had
default values instead of those configured!