Discard invalid comet requests when checking cometForHost.
Invalid comet requests are requests that throw an exception when we try
to access their hostname. This is known to happen in Jetty, and
manifests in slightly different ways across containers, when we attempt
to read data from a request that the container is no longer willing to
service. These requests are immediately discarded using the existing
makeCometBreakoutDecision code.