* Fixes 805 - snippet Msgs.ShowAll should not be a session var
By being a SessionVar, it meant that if you had two pages, one with showAll=true and the second
page with showAll=false (or omitted), messages that correspond to an ID would show or not
depending on which page was loaded last (setting the SessionVar to true or false).

Now we use a RequestVar and all is well.