Add cometUpdateExceptionHandler LiftRule.
Not implemented yet, but this rule specifies a Box[JsCmd]. If Full, the
contained JsCmd provides the code to put inside the catch of a try/catch
block that is wrapped around comet updates. If Empty, the old behavior
of not wrapping comet updates in try/catch at all is used.