Add LiftRules noCometSessionCmd/noAjaxSessionCmd.
These are the JsCmds that will be sent to the client when a session is
missing for a comet or ajax request, respectively. For comets, no comet
session is a state where a comet request comes in with no associated
session OR with a session that has no associated CometActors (typically
this happens on server reload).