Include a version on the end of the Ajax GUID.
AJAX requests currently carry a GUID. We now add a single-character
version indicator. This version increments for distinct AJAX requests.
In particular, it does NOT increment during AJAX retries, so that a
retry of an existing request can be identified as being part of the same
attempt on the server.

Not all AJAX requests carry this version identifier. In particular, Lift
GC requests do not carry a version identifier. This is because these
are going to be handled in a streamlined handler. If a Lift GC request
doesn't make it through, we can retry it as many times as we want and
it'll just remark stuff in the session.

liftAjax.addPageNameAndVersion appends both the page GUID and the
version number. The version number is encoded in base-36.