AJAX version-based deduplication.
The meat of the deal. Based on the AJAX version appended to the request
GUID, we determine whether we've already seen this request. If so, we
wait for the original request to complete before returning the resulting
value.  If we already completed the request before, we return the same
answer without re-running the associated parameters. AJAX requests that
need to wait are put into continuations if available.