Implement the meat of AJAX deduplication.
AJAX requests now come in with a two-part version number, one a sequence
number for the ajax request's count in the overall list of requests sent
by the client, and one a number indicating how many other requests are
queued up on the client.

If this is the first request seen with its sequence number, we record
the request in the session and run regular AJAX request handling. When
that request handling is complete, it satisfies the future that is
recorded in the session. Subsequent requests for a given sequence number
wait on the future from the first request up to the ajax post timeout,
then fail.

When we get a request coming in with a 0 count for pending requests on
the client, we clear out all other pending requests in the session's
list, since that means none of them will be getting a chance to report
their responses again.