Rewrite SoftReferenceCache.apply for type safety.
We were using a tuple assignment, which converts into a Scala pattern
match that doesn't carry particularly nice type safety guarantees. We
now instead assign to a typed tuple and then unpack the tuple with our
own match statement, with a clarifying comment on the original
assignment to indicate what each part of the tuple means.