Add centralized handling for parameters not mapped to Lift functions.
We add a LiftRules entry, handleUnmappedParameter, that is a FactoryMaker that
vends a function to handle any parmaeter not mapped to a Lift function. By
default, the function checks if the parameter name starts with F, and logs a
warning if so. This is to do a quick, high-speed check as to whether or not the
parameter looks like a Lift function binding.