Add helpers for detecting IE10/11.
I know we were all hoping that by IE10 they'd be up to speed with the
rest of the world, but I still remember some support issues, so I'm
adding a detector for it in the event server side detection is needed
for some people. Ditto for 11. Here's hoping these will be the last
versions of IE that ever need such detection.