Simplify Loc.doesMatch_?
As I was reading over this code to look at this issue, we were calling
Link.isDefinedAt to determine whether or not to call Link.appy.
Link.apply calls Link.isDefinedAt to determine whether or not to throw
an exception.

This seemed redundant in this particular case, so I've reduced this
method to a simple boolean expression that should have the same effect.