Add and implement the MatchWithoutCurrentValue LocParam.
The MatchWithoutCurrentValue LocParam overrides Lift's default behavior
of considering an Empty currentValue a non-match for a Loc. This default
action is desirable in most circumstances, but occasionally we want to
allow client code to define some custom behavior in the event
currentValue comes up Empty.

Using this new LocParam in combination with IfValue would allow them to
do so.