Define a get method on Box and immediately deprecate it.
We do this so that get no longer triggers an implicit conversion
to Option, and triggers a deprecation warning. In the future,
get will remain defined but will throw an exception and be
defined as returning an internal vestigial type to break
compilation.

Those who still want to use .get can call .toOption.get.