Fix ne comparison warnings in lift-mapper tests.
In MapperSpec, because the providers list was inferred to
have type List[DbProviders.H2MemoryProvider], comparisons
to avoid testing certain things against MySql and other
providers were causing warnings.

We fix this by being explicit about the type of the providers list.
This is important because the spec is structured so that you
can just add other providers if you have them set up properly
and test mapper as needed against those providers.