Turn ToCssBindPromoter to implicit classes.
Before, we had a couple of implicit conversions from String to
ToCssBindPromoter, which was a final case class. We now
have a CssBindImplicits trait that defines a CssBindPromoter
class and two implicit classes that extend it, one for strings
and one for CssSelector instances.

These implicits are mixed into HtmlHelpers so they are
available, as before, when Helpers._ is imported.