Add restartComet() for restarting comet cycle.
restartComet is meant for cases where, for example, a new
comet is added to toWatch and we want its presence
reflected immediately. Without it, we would have to wait for
the current comet request to finish, which could take up to
90s if the old comets had no further updates, and could
bottle up updates from the new comet for that long, because
the new comet wouldn’t be able to push its contents through
the old connection.