Add CometId case class.
CometId describes a comet uniquely in the session, and
consists of the comet type and the (possibly Empty) comet
name. The cometSetup SessionVar and nasyncComponents
ConcurrentHashMap both use CometId for indexing.

Not all places that interact with these have been updated,
as there are chunks of LiftSession that are about to get
extensively rewritten; as such, this doesn’t currently
compile.