Significantly reorganize comet find/setup stuff.
Things are split up a little more granularly. We also have a
more uniform and descriptive API (findComet only finds
existing comets, while findOrBuildComet is used to build a
comet if it wasn’t found). The API is still internal to Lift’s
packages.

This still doesn’t quite compile, as it breaks the Comet
snippet.