Rework Comet snippet.
In particular, the snippet now exposes a
Comet.containerForCometActor method that can be used by
an arbitrary third party consumer to produce the container for
a given comet actor when a handle to the LiftCometActor
object already exists.

It is, of course, still possible to materialize a CometActor using
regular lift:comet snippet.