Track comets to add to LiftMerge using RequestVar.
Before, comet tags included a lift:when attribute that was
used to extract comet versions and such. We now track
these when S.findOrCreateComet is called, and in LiftMerge
we use this list instead of looking for the marker attribute.

Because all invocations of the Comet snippet go through
S.findOrCreateComet, these are automatically tracked;
however, things added via S.addCometAtEnd are currently
lost, and we’re not handling AJAX send-down yet.