Extend FlatSpec instead of mixing it in.
We do this in a way that will work across both 2.9.x and
2.11 versions of scalatest, which means we have to treat
FlatSpec as a class. Mixing in FlatSpecLike isn’t an option
because the trait doesn’t exist on the 2.9.x version of
scalatest.

To do this properly, we have to instantiate a LineTokenizer
directly in the class rather than being able to extend it in our
test class, since you can’t extend two classes.