Add nasty 2.9.2 compatibility hack.
Scala 2.10 introduced the collection.concurrent.Map
package and class that supersedes ConcurrentMap, and
then 2.11 removed ConcurrentMap altogether.

2.10 also introduced JavaConversions for the new type, and
2.11 removed the old ones. To properly compile in 2.9, we
need a thin layer that provides the new APIs in 2.9 and
transparently calls the old stuff.

Also mucks about thoroughly with the project setup so that
we have a pre-2.11 project that does all the nasty 2.9 slicing
and dicing and a regular project that does everything
2.11-style.