updateListeners(Any)->sendListenersMessage.
We do this because there is an updateListeners overload that takes a
list of actors to update, which is optional. Having an updateListeners
overload that takes an Any makes it really easy to accidentally call the
wrong updateListeners method by having the wrong type of List.
Additionally, it's particularly confusing to have the two methods named
the same thing when they behave differently.

We name the method that sends all listeners a message that is passed in
`sendListenersMessage` and leave `updateListeners` for sending all
listeners the message generated by `createUpdate`.