updateListeners(Any)->sendListenersMessage.
We do this because it's particularly confusing to have the two methods
named the same thing when they behave differently.

We name the method that sends all listeners a message that is passed in
`sendListenersMessage` and leave `updateListeners` for sending all
listeners the message generated by `createUpdate`.

The old updateListeners(Any) method is deprecated for the 2.6 cycle.