Do basic example extraction from dexy using Ruby script.
Not doing anything with the extracted examples yet, but this gives us the
starting point from which we can look at how we can read in a file, find the
examples in it, and run them. Ultimately, we're going to want to do this all
in Scala with CSS selector transforms, because hey, why not.