Fix dexy invocation of selector example extractor.
We now want to be passed the directory that the built HTML
is in, and we also have to compute it.

Additionally, we make use of the DEXY_ROOT environment
variable that dexy sets for us to know the project’s root
directory for proper invocation of sbt (we were computing this
in a somewhat brittle way before).