Rework LazyLoad comet.
We move the AsyncRenderComet to its own file. We also
move the actual render function management into the
AsyncRenderComet companion, including handling deferred
function building.

LazyLoad slims down a little, dealing with invoking the
async render setup in AsyncRenderComet and rendering
the placeholder.

We also only apply a div wrapper to the default loading
template, letting a custom template stand on its own, though
we apply an id to it for the purposes of replacing it with the
rendered contents after they are sent down.