Add buildDeferredFunction for (A)=>T.
LiftSession’s buildDeferredFunction only existed for a no-arg
function. We now have a version that builds a deferred
function of one argument; invoking that function with an
argument will restore the session/request context and then
execute the original function with the passed argument.

Also simplified our restoration of context in both versions of
buildDeferredFunction.