Fix lift.logError binding.
We were tying settings.logError directly to lift.logError, but
this didn’t allow for the options passed to init to override the
implementation because lift.logError is bound before the init
function runs. We now delegate to settings.logError using
apply, ensuring we’re always seeing the latest version of it.