Extract on* attributes to page JavaScript.
We now pull out all on* attributes, moving them instead to
onEvent invocations in the page JavaScript. Where
necessary, we mark the element in question with a unique id
so we can address it in the onEvent invocation.

This lets us remove all inline JavaScript from the page,
preparing us to be able to set strong Content-Security-Policy
settings out of the box.