Enrich LazyLoad snippet with some render tweaks.
There is now an overload that takes a renderer function to which we hand the
placeholder id that we generate, and which is charged in turn with doing the
final rendering. It expectes the caller to wrap the passed function in whatever
deferred handling may be needed.

The main render method also now has an optional placeholderTemplate parameter
so that when invoked from Scala it can be given a NodeSeq to use for the
placeholder.