Fix issue with new password verification in ProtoUser.
When we converted from the original bind strategy, I forgot that the old
bind strategy evaluated things eagerly, while CSS selector transforms
call by name. This meant that the old bind strategy, when it bound both
new password fields (the new password and repeat new password fields) to
a single SHtml.password_* invocation, actually bound both fields to the
same password input with the same GUID.

In the CSS selector transform port, the result was instead two fields
with two different GUIDs, which broke the expected behavior of the
password field (it was supposed to generate a List of two passwords,
which could then be checked for equality).

To fix this, we now create the single field ahead of time and pass that
to the selector transform so that the same field gets used twice instead
of a different field being generated each time.