Don't lose jsToAppend during comet AJAX request.
When dealing with an AJAX request in a comet context, we
were partial-updating jsToAppend at the beginning of the
request, but any updates that might have been made to it
during the request processing were lost. We now send the
partial update message after the request functions have run,
so that any JS they’ve appended also gets sent down to the
client.