Avoid duplicate requests for comet-sourced comets.
If a comet was sent down from within a comet context, we
would force a comet request reset. On the client, this would
schedule a new comet request and attempt to abort the old
one; however, there is no way to abort the old one, since its
success handler is what is kicking off this request reset.

The end result is that the request’s success handler finished
running, which would also schedule a new comet request,
and we’d get two comet requests running in parallel and
walking all over each other.

When generating the comet commands, we now check to
see if we’re executing in a comet context, and if we are we
ask the client not to reset the comet cycle, knowing that the
response we’re currently sending will immediately kick off a
new request once it’s done executing.