Add ability to link to other Scaladocs in ours.
We do two things: turn on autoAPIMappings, which
automatically lets us link to Scaladocs for modules that set
their public URIs, and explicitly add the mapping for the
scala-xml module. The addition of the scala-xml module is
also reasonably generic, so we can add more module-specific
URIs fairly easily in the future as needed.