Fix two issues with LiftMerge id extraction.
When we were detecting the id for use by event handlers, we
were doing two things wrong:
 - We were failing to pass on any existing event attributes that
   we’d extracted along the way, so if there were any they
   were lost.
 - We were failing to append the fixed attributes to the current
   attribute, so we were losing the id attribute altogether.

This meant that if an element had an id, we always lost that
id in the process of trying to reuse it for event attributes. We
also only kept the event attributes for attributes before the id,
as the ones after were lost while we processed the id.