Support extracting javascript: actions and hrefs.
javascript: and javascript:// are used in a couple of places in
Lift, and can be used in client code. We look for these in
action and href elements and extract them into proper event
handlers with return false; appended (since return false is
implied in these situations).

We also clarify some variable names and comments along
the way.