Revert "Revert "Merge branch 'arkadius-clean-comb-parser-helpers-spec-fix'""
This reverts commit bf25bf6887f7c6b10dadd84849c17454e9d0b6a4. The
original revert was there because I cherry-picked a commit I wanted
on master onto the a branch off of this one, and then merged it
into master before this one was ready to go (shoulda opened a PR...).
This commit reverts the revert on this branch, so that when we merge
this branch to master its changes will take hold (otherwise the first
revert would make the original changes be lost). Whew!