Ensure uniqueness of comet requests.
We now track a "current count" for a comet request, and any calls to start a
request carry the then-current count. Once a comet request is fired, the count
is incremented, ensuring that even if a new comet request is scheduled multiple
times, only one will run at any given moment.

This is only the new part of the guard. The other part of the guard is that we
track the actual current comet request and abort it if a comet request restart
is needed. However, an AJAX request that was just begun cannot be aborted, so
if multiple restarts are scheduled in one tick for whatever reason, we end up
with multiple requests--this fix guards against that scneario.