Fix an issue where we could lose event attributes.
This particularly manifested when the attribute whose href we wanted to fix
(e.g., the `href` attribute on an `a` element or the `action` attribute on a
`form` element) also had an event handler (e.g. `onclick`). The code that fixed
the href failed to pass on the event handlers that had been seen so far to the
attribute-fixing chain, so they got lost before they could be applied.