Don't assume postPageFunctions are Some.
When generating post-page JavaScript, we were assuming that because our first
access of the post-page functions was non-empty, all of them would be. This led
to occasional .get-related NullPointerExceptions. We now allow for later calls
to the post-page functions to be None, and simply assume we have no more
functions to fetch once we get a None for them.