Only build master and lift_26 on push.
Without this specifier Travis will attempt to execute two jobs for each
PR: the PR-merged job and the push job. The former is really the most
useful of the two. Also, without this specifier Travis will execute a
build for every push to every branch. Both of these conditions mean
we're making Travis do more work than is really useful.

With this change we should only be running builds on:
  - Push to master to lift_26
  - Opening or push to open PR.