Simplify bufRenderArr.
It looks like bufRenderArr was doing a check for null that we didn't
really need, since bufRender already handles null correctly. We also
were doing a match when we didn't really need one. I've replaced the
match with a simpler if-statement that should result in less code to
execute along this critical path.