Add clearAfterRead parameter to jsToAppend.
This will change the way you have to invoke jsToAppend to require
parens, but allows for jsToAppend to clear the appendable JS after
having assembled it to be returned. This ensures Lift’s internals won’t
double-evaluate jsToAppend in weird ways.

It may be worth making this private[http] and exposing a public version
that takes no parameters and sets clearAfterReading to false, we’ll see.