Add back Printer methods.
This means that any code that relies on using the code style
  compact(render(value))
will still compile and work as expected in client code, but will now
generate a deprecation warning.

This would, however, break code that relies on the scala.text.Document
format. That may be undesireable, but my current reasoning for why that
is O.K. is that scala.text.Document has been deprecated in Scala since
2.11.0 so, in theory, anyone relying on that format has had time to
implement work arounds.

I'll throw that reasoning out on the ML. Others may insist we support
scala.text.Document somehow for another release.