Only swallow MappingException for JNothing/JNull.
Previously, if a field was considered optional, we would eat any
conversion exceptions that occured while trying to understand that
field. This meant that if something was an Option and we couldn't
understand what was at that key, we'd just give you a None.

Now, if something is an Option and nothing is provided for that key then
we still give you a None. But if something *is* provided for that key
and it doesn't make sense to us, we're going to give you a
MappingException because you probably are expecting something to be
there and that thing is of the wrong format.