Implement the travis.sh build script.
This build script for travis will check whether or not we're currently
in a Travis PR build. If not, it'll check the branch. If we're on master
or lift_26, we'll trigger a SNAPSHOT image deploy.

Of course, we don't yet have the credentials in travis yet so this won't
_quite_ work how we want it for the moment. But give me more than one
commit to get that sorted out!

:construction_worker: