Switch builds to use JDK 8.
There was a discussion on the Lift list about switching to building with
Java 8 for Lift 3 so we could support Scala 2.12 when it comes out. I
don't remember if there was a strong consensus against this, but I do
want to go ahead and start moving that direction, so we're going to
switch the Travis build over to that.