Fix BaseParsers ws parser to not match newlines.
In lift-markdown, BaseParsers was using a regex containing `\v` to indicate a
vertical tab. As of Java 8, `\v` refers to a character class that contains all
vertical whitespace, including newlines. We now explicitly use the unicode
value for vertical tab in the regex, fixing the failing test in Java 8 and
generally making things work as expected.