Add CssHelpersSpec and fix bugs in CssUrlPrefixer.
CssHelpersSpec is properly capitalized despite the fact that it's hard for us
to recapitalize CSSHelper without breaking things; it reflects the proper
capitalization approach in Lift.

It also tests several common scenarios, and verifies that the URL prefixer does
not change the quoting type that is used in the source CSS. This is important
for URLs that may contain unescaped single quotes inside double quotes or vice
versa. Additionally, we verify support for commas and other special characters.
This yields fixes for comma support and percent sign support.