Add custom JSON character escaping, and defaults for JS.
Until now, we've had a hardcoded set of characters that would always be escaped
in JSON; these were never modifiable. We now plug in to the new RenderSettings
case class to add the ability to specify sets of characters that should be
escaped.

Additionally, we add a default set of such characters that corresponds to the
characters that must be escaped when JSON is evaluated directly by a JavaScript
engine rather than by a valid JSON parser (as when using JSON-P).

Lastly, we provide two default RenderSettings instances, prettyJs and
compactJs, that encapsulate the appropriate escaping with pretty or compact
rendering.