Fix hasOwnProperty use in IE.
There was code that referenced a nonexistent hasOwnProperty variable, which we
now introduce, and we change the one other usage of hasOwnProperty to also use
this variable. This should generally be IE-safe---the core issue is that IE 8
does not have hasOwnProperty on native objects like Array, which can cause
things to go boom boom.