Drop complete_? checks in CanResolveAsyncSpecs.
We were using that to fail if the spec hadn't sastified its
result future with the value received from the async resolution,
but it looks like those satisfactions happen async, so we can't
be sure that the future will be immediately complete once we
hit that line of code. Instead we just let `get` block until we get
the proper resolution, which should still be very quick.