Add WebJars support
* Added WebJars snippet that calculates the full path to an asset.
* Added WebJarServer to serve the assets if not running in a Servlet 3 container
* Moved lift.js, jlift.js, liftExtCore.js, and liftYUI.js to *src/main/assets/js*. They are now included in the webkit JAR and can be served as any other WebJar asset can. LiftMerge was updated to output a WebJar compatible script tag. Backwards compatibility for serving these JARs using */classpath* is provided, with a warning being logged. See HTTPProvider.preBoot for details.

All other files in *toserve* will be served as before.

In order to do this I had to remove the YUICompressor plugin because the version of Rhino it uses conflicts with the one sbt-web uses. So, I added all of the minified versions of files in *toserve* that YUICompressor creates directly to the repo so they don't have to be compressed during the build anymore.