Add HtmlNormalizer.
This is a singleton that holds functions related to normalizing HTML
URLs as well as extracting JS event handlers from HTML. It’s designed
for use in more than just LiftMerge, and is wired into LiftSession.

We also deprecate LiftSession.fixHtml in favor of normalizeHtml, and
add support in JsCommands for extracting event handlers from HTML
rendered in JS responses.