Turn off prod Content-Security-Policy enforcement by default.
We introduce flags to toggle CSP enforcement in modes other than dev, and
default to false. This is to smooth transition to Lift 3.0.0 as discussed on
the mailing list; 3.1.0 will shift to the final secure default of enforcing in
production.

Additionally, in the process, we rework where we do mode checking---we now
check the run mode in SecurityRules and pass the category-specific header
generators final information about whether enforcement and logging are enabled.