Rework how prop providers and interpolators work a little.
We define two independent types (PropProvider and InterpolationValues), even
though they ultimately equate to Map[String,String]. We also rename the
associated functions so that they agree with the type names. Finally, we only
provide helpers for adding to the prop providers and interpolation values,
and provide one updater function in case you want to do more complex mutation.

We also lock down the interpolator and prop lists on first access via an
intermediary lazy val so that we're not having to manage concurrently mutable
state.