Make providers respect whereToLook.
Due to how we'd structured provider loading, we eagerly looked up props when
the Props singleton was first instantiated. This prevented the customization
point that is updating `whereToLook` from working.

We now indirect providers access a bit so as to lazy-load providers when we
first access a property, and then we can respect whereToLook and look things up
as the user wants them.

Also added a test verifying this behavior.