Append extracted events to returned JS in fixHtmlAndJs.
Before, we were using a global appendJs for these events. However, fixHtmlAndJs
already provides a mechanism for returning additional JS. We now use it.

The previous behavior would result in extracted events being appended to the
response (page or AJAX) even if the JsCmd that invoked fixHtmlAndJs was never
actually sent down to the client. This behavior correctly makes the extracted
event JS part of the JsCmd, so that it can be sent down only with that JsCmd.