Reduce segment churn in Buffer.
We now only add new segments to the buffer’s list of segments when
absolutely necessary. In particular, marking the buffer is an
indication that we’re interested in reading between the current segment
and whenever `substring` is next called. If the buffer is unmarked, we
always reuse the same segment. We don’t release a segment until the
buffer is no longer in use. And any segments beyond the first that we
allocate are reused for future marking operations that require more
than one segment.