Rework `unquote` for better data sharing.
`Buffer` now provides a unique StringBuilder that’s used during
`unquote`. We also rework `substring` to allow for emitting the
substring between the marked point and the current point directly into
the builder, which reduces the number of array copies we need to do
before materializing the full string with escape characters handled.

We also rework the next-character match for escaped characters to be
optimized to a switch by the compiler.

Lastly, we rework the way hex characters are parsed to avoid
`Integer.parseInt` and an extra string creation.