Build up arrays and objects with ListBuffer.
Before, we were building up JObject fields and JArray entries in
immutable data structures and in reverse, so we had to reverse them
when they were completed. We now temporarily use a ListBuffer, which
allows constant-time mutable append and cheap conversion to our final
List data type, which we do once the array or object is done being
built up.