Optimize json parseValue.
Before we would instantiate a StringBuilder when we wanted to parse a
value and append characters to it as desired. Now, we use the buffer’s
marking functionality to mark the beginning of the value and then
simply increment the current position in the buffer until we’re ready
to wrap up. Then we do a single substring() operation that equates to a
single array copy, and fork that over to double or int value conversion.

We also stop using Character.isDigit, which is both slow and incorrect
(only 0-9 are accepted digits for JSON numbers---other Unicode digits
are not), and optimize the three cases (decimal/exponent, digit/sign,
everything else) to a three-part switch.