Rip out bad-double parsing bug fix.
A version of JVM 1.6 had an issue where a particular double construction could
trigger the JVM to hang. We had a workaround to that issue that required
parsing doubles with BigDecimal and such. That bug is long since fixed in JVMs,
both with a patch on 1.6 and forward into the newer versions. Additionally, Lift
3 is no longer built for Java 6. As a result, we rip out BigDecimal and go back
to the simpler String.toDouble construcation.