Request/session - aware futures
Lift 3 allows to bind Future / LAFuture instance to template elements easily. Still, when Future gets transformed with one of its map/flatMap/etc methods or when Future body gets executed we don't have an access to current Req or LiftSession. This issue seems to be recurring between various projects using Lift framework. This commit brings request and session-aware futures. It allows to create such a type of LAFuture or decorate Scala's Future instance so that Req and LifSession parameters can be accessed from it.