Mark JValue functor composition pending until fixed.
Once we implemented proper JValue=>JValue generation, it became evident
that JValue’s map does not actually adhere to this property, thus we
mark it pending until fixed. Actually fixing the issue would require
changing a pretty core aspect of how map behaves, so it’ll have to wait
for a future release.