Move segment size example to be run sequentially after others.
The segment size spec changes the segment size to make sure that parsing
still works. However, the segment size is now used when calculating the
size of array to allocate when taking substrings of a buffer. So, while
running an entire parse with the same segment size works fine, changing
the segment size mid-parse does not. Because the segment size example
ran concurrently with the other parsing examples, it could trigger
failures in those examples and in itself.

We now run it sequentially after all the others have finished, so that
it can do its testing in isolation. The result... Works!