Fix cross-compile issues on master.
Scala 2.12 added a few methods to the Future interface. After trying a
number of clever work-arounds using macros, the best way to smooth the
gap that I could find was to throw an UnsupportedOperationException if
either of these new methods are invoked.

This is the "least bad" solution I could come up with given the reality
of the situation. I would be very greatful for any other options as
throwing these exceptions is kind of lame.