Change how we test segment size changes.
We were testing segment size changes by using mutable state on a global
Segments singleton. With the new parsing approach, which aggressively
uses mutability to increase performance, this was causing derivative
concurrency failures in many places.

We now change the way to test different segment sizes by putting an
interface in front of the segment pool and allowing it to be customized
on a per-Buffer basis, and making the spec instantiate the Buffer with
a different segment pool while testing.

This functionality is currently still entirely private in the json
package, but can be exposed in the future if we deem it useful. For
now, its sole purpose is to confirm that buffer size can be changed
without adverse effects.