Correctly map tuples to the HCol concept.
The mapping engine here determines what we actually try to instantiate
when we see incoming JSON and we're trying to create a concrete series
of objects through extraction.

Here, we indicate to the rest of the system that if we're looking for a
tuple in the instantiated type we'll expect the JSON representation to
be an array of some sort.