Retain the ability to interpret old tuples.
Before my changes, tuples were serialized as follows:

  {"_1": "value", "_2": "value2"}

After my changes, tuples are serialized like this:

  ["value", "value2"]

But we still will stumble on a lot of tuples that have been written out
to some permanent storage in the old format.

Now, if we expect there to be a tuple in a particular object but we
found a JObject representation we will convert it to a JArray under the
hood and recursively re-invoke the newTuple function.

This will only apply if all the keys in the object start with an
underscore, which should both be sufficiently cheap and sufficiently
idicitive of a tuple since we're expecting a tuple in the field we're
trying to inflate anyway.