`fixHtmlFunc` never extracts inline JS.
Before, `fixHtmlFunc` would attempt to extract inline JS if that behavior
was enabled in `LiftRules.extractInlineJavaScript`. Unfortunately this
didn't work too well in many uses of `JsExp`, due to the fact that the
extracted JavaScript would run before the HTML that it was extracted
from would be attached to the DOM.

We need a longer-term solution, but for now we introduce a pathway for
callers of `normalizeHtmlAndEventHandlers` to force `extractInlineJavaScript`
functionality on or off, and we use it to turn it off always for `fixHtmlFunc`.

We also add a spec for `HtmlFixer` that verifies this behavior.